improved biome worldgen

+++ Engine/AABB.cpp
@@ -7,11 +7,11 @@ bool AABB::IntersectsRay(const Vector3& origin, const Vector3& direction) {
 	Vector3 tMin = (min - origin) * invDir;
 	Vector3 tMax = (max - origin) * invDir;
 
-	Vector3 tEnter = Vector3(fmin(tMin.x, tMax.x), fmin(tMin.y, tMax.y), fmin(tMin.z, tMax.z));
-	Vector3 tExit = Vector3(fmax(tMin.x, tMax.x), fmax(tMin.y, tMax.y), fmax(tMin.z, tMax.z));
+	Vector3 tEnter = Vector3((float)fmin(tMin.x, tMax.x), (float)fmin(tMin.y, tMax.y), (float)fmin(tMin.z, tMax.z));
+	Vector3 tExit = Vector3((float)fmax(tMin.x, tMax.x), (float)fmax(tMin.y, tMax.y), (float)fmax(tMin.z, tMax.z));
 
-	float tEnterMax = fmax(fmax(tEnter.x, tEnter.y), tEnter.z);
-	float tExitMin = fmin(fmin(tExit.x, tExit.y), tExit.z);
+	float tEnterMax = (float)fmax(fmax(tEnter.x, tEnter.y), tEnter.z);
+	float tExitMin = (float)fmin(fmin(tExit.x, tExit.y), tExit.z);
 
 	return tEnterMax < tExitMin;
 }
